requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
idna==3.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
